+++ modules/eks/provider.tf
@@ -17,7 +17,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 6.0"
+      version = "~> 5.95"
     }
     null = {
       source  = "hashicorp/null"