+++ modules/iam_irsa/outputs.tf
@@ -14,23 +14,7 @@
 
 # modules/iam_irsa/outputs.tf
 
-output "alb_irsa_role_arn" {
-  description = "IRSA role ARN for the AWS Load Balancer Controller"
-  value       = aws_iam_role.alb_irsa.arn
-}
-
-output "aws_iam_openid_connect_provider" {
-  description = "IAM OIDC provider for the EKS cluster"
-  value       = aws_iam_openid_connect_provider.eks_oidc_provider
-}
-
-output "oidc_provider_arn" {
-  description = "ARN of the OIDC provider"
-  value       = aws_iam_openid_connect_provider.eks_oidc_provider.arn
-}
-
-
-output "oidc_issuer_host" {
-  description = "OIDC issuer host for the EKS cluster"
-  value       = replace(var.oidc_issuer_url, "https://", "")
-}
+# output "efs_csi_driver_role_arn" {
+#   description = "IAM role ARN for the EFS CSI Driver"
+#   value       = module.efs_irsa.iam_role_arn
+# }