+++ modules/iam_irsa/provider.tf
@@ -15,7 +15,7 @@
 terraform {
   required_version = ">= 1.5.0"
   required_providers {
-    aws = { source = "hashicorp/aws", version = "~> 6.0" }
+    aws = { source = "hashicorp/aws", version = "~> 5.95" }
     tls = { source = "hashicorp/tls", version = "~> 4.0" }
   }
 }