+++ modules/vpc/main.tf
@@ -24,7 +24,7 @@
 */
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "~> 6.4"
+  version = "~> 5.8"
 
   name = "${var.identifier}-${var.environment}-vpc"
   cidr = "10.0.0.0/16"
@@ -47,11 +47,14 @@ module "vpc" {
 
   # Helpful tags on subnets
   public_subnet_tags = {
-    "kubernetes.io/role/elb" = "1"
+    "kubernetes.io/role/elb"                              = "1"
+    "kubernetes.io/cluster/${var.identifier}-eks-cluster" = "shared"
   }
   private_subnet_tags = {
-    "kubernetes.io/role/internal-elb" = "1"
+    "kubernetes.io/role/internal-elb"                     = "1"
+    "kubernetes.io/cluster/${var.identifier}-eks-cluster" = "shared"
   }
+
 
   tags = var.common_tags
 }